requirements.txt
Flask==2.0.2
Flask-Injector==0.13.0
SQLAlchemy==1.4.27
Flask-SQLAlchemy==2.5.1
Flask-Injector==0.13.0
PyMySQL==1.0.2